drafted API for Scheduler
This API will most likely be what the master front-end uses to manage the overview shown to the user. As such, I wanted to make it as simple and powerful as possible.

Right now, the API does not provide any way to remove LiveMasters from the scheduler. I cannot figure out a good way to provide that functionality while maintaining a useful WaitMasters() function. Later, after I start implementing the front-end, I may be more capable of thinking of something satisfactory.