use more natural template indentation
Before, I was concerned with making the output HTML properly aligned. Now, I have pretty much decided that that is impractical (since some of the re-usable templates might be used in different indentation contexts). Instead, I am indenting the files in a way that seems "natural" when reading the template code.