fixed confusing date formatting bug.
Before now, the day-of-month field of a formatted date would be erroneously set to the hour (mod 12) of the timestamp. For instance, Jan 13 2016 at 16:45 would be formatted to Jan 4 ...

This took a while to notice because the date on a given timestamp is not usually as relevant as the time itself, at least not when jobs take 15 minutes or so to run.